+++ tsconfig.json
@@ -26,7 +26,7 @@
     "checkJs": false,
 
     "baseUrl": ".",
-    "outDir": "lib",
+    "outDir": "dist",
     "lib": ["es2017", "dom"],
 
     "typeRoots": ["./node_modules/@types"]
